added eslint ignore comments
