+++ packages/map/src/CdcMap.tsx
@@ -1,5 +1,4 @@
-import React, { useState, useEffect, useRef, useCallback, useReducer } from 'react'
-import * as d3 from 'd3'
+import { useEffect, useRef, useCallback, useReducer } from 'react'
 
 // IE11
 import 'whatwg-fetch'
@@ -35,7 +34,6 @@ import { DataTransform } from '@cdc/core/helpers/DataTransform'
 import MediaControls from '@cdc/core/components/MediaControls'
 import fetchRemoteData from '@cdc/core/helpers/fetchRemoteData'
 import getViewport from '@cdc/core/helpers/getViewport'
-import Loading from '@cdc/core/components/Loading'
 import numberFromString from '@cdc/core/helpers/numberFromString'
 import DataTable from '@cdc/core/components/DataTable' // Future: Lazy
 
@@ -54,6 +52,8 @@ import UsaRegionMap from './components/UsaRegionMap' // Future: Lazy
 import WorldMap from './components/WorldMap' // Future: Lazy
 import useTooltip from './hooks/useTooltip'
 import mapReducer from './store/map.reducer'
+import { hashObj } from './helpers/hashObj'
+import { generateRuntimeLegend } from './helpers/generateRuntimeLegend'
 
 // Data props
 const stateKeys = Object.keys(supportedStates)
@@ -70,38 +70,7 @@ const generateColorsArray = (color = '#000000', special = false) => {
   return [color, hoverColor, colorObj.darken(0.3).hex()]
 }
 
-const hashObj = row => {
-  try {
-    if (!row) throw new Error('No row supplied to hashObj')
-
-    let str = JSON.stringify(row)
-    let hash = 0
-
-    if (str.length === 0) return hash
-
-    for (let i = 0; i < str.length; i++) {
-      let char = str.charCodeAt(i)
-      hash = (hash << 5) - hash + char
-      hash = hash & hash
-    }
-
-    return hash
-  } catch (e) {
-    console.error('COVE: ', e) // eslint-disable-line
-  }
-}
-
-const indexOfIgnoreType = (arr, item) => {
-  for (let i = 0; i < arr.length; i++) {
-    if (item === arr[i]) {
-      return i
-    }
-  }
-  return -1
-}
-
-// returns string[]
-const getUniqueValues = (data, columnName) => {
+const getUniqueValues = (data, columnName): string[] => {
   let result = {}
 
   for (let i = 0; i < data.length; i++) {
@@ -125,12 +94,12 @@ type CdcMapProperties = {
   isEditor?: boolean
   hostname: string
   link?: string
-  navigationHandler?: any
+  navigationHandler?: Function
   isDebug?: boolean
   logo?: string
-  setConfig?: () => any
-  setSharedFilter?: () => any
-  setSharedFilterValue?: () => any
+  setConfig?: Function
+  setSharedFilter?: Function
+  setSharedFilterValue?: Function
 }
 
 const CdcMap = (props: CdcMapProperties) => {
@@ -162,9 +131,10 @@ const CdcMap = (props: CdcMapProperties) => {
     loading: true,
     modal: null,
     position: configObj?.mapPosition ?? initialState.mapPosition,
-    runtimeData: [],
+    runtimeData: {},
     runtimeFilters: [],
-    runtimeLegend: []
+    runtimeLegend: [],
+    runtime: {}
   }
 
   const [state, dispatch] = useReducer(mapReducer, mapInitialState)
@@ -207,7 +177,7 @@ const CdcMap = (props: CdcMapProperties) => {
     dispatch({ type: 'SET_FILTERED_COUNTRY_CODE', payload })
   }
 
-  const setRuntimeData = payload => {
+  const setRuntimeData = (payload: Record<string, Object>) => {
     dispatch({ type: 'SET_RUNTIME_DATA', payload })
   }
 
@@ -221,6 +191,10 @@ const CdcMap = (props: CdcMapProperties) => {
 
   const setDimensions = payload => {
     dispatch({ type: 'SET_DIMENSIONS', payload })
+  }
+
+  const setErrorMessage = (payload: string) => {
+    dispatch({ type: 'SET_ERROR_MESSAGE', payload })
   }
 
   const transform = new DataTransform()
@@ -255,7 +229,7 @@ const CdcMap = (props: CdcMapProperties) => {
         const tmpData = {
           [filteredCountryCode]: filteredCountryObj
         }
-        dispatch({ type: 'SET_RUNTIME_DATA', payload: tmpData })
+        setRuntimeData(tmpData)
       }
     }, 100)
   }, [filteredCountryCode]) // eslint-disable-line
@@ -300,7 +274,7 @@ const CdcMap = (props: CdcMapProperties) => {
   // Tag each row with a UID. Helps with filtering/placing geos. Not enumerable so doesn't show up in loops/console logs except when directly addressed ex row.uid
   // We are mutating state in place here (depending on where called) - but it's okay, this isn't used for rerender
   // eslint-disable-next-line
-  const addUIDs = useCallback((obj, fromColumn) => {
+  const addUIDs = (obj, fromColumn) => {
     obj.data.forEach(row => {
       let uid = null
 
@@ -374,500 +348,15 @@ const CdcMap = (props: CdcMapProperties) => {
     })
 
     obj.data.fromColumn = fromColumn
-  })
+  }
 
   // eslint-disable-next-line
-  const generateRuntimeLegend = useCallback((obj, runtimeData, hash) => {
-    const newLegendMemo = new Map() // Reset memoization
-    let primaryCol = obj.columns.primary.name,
-      isBubble = obj.general.type === 'bubble',
-      categoricalCol = obj.columns.categorical ? obj.columns.categorical.name : undefined,
-      type = obj.legend.type,
-      number = obj.legend.numberOfItems,
-      result = []
-
-    // Add a hash for what we're working from if passed
-    if (hash) {
-      result.fromHash = hash
-    }
-
-    result.runtimeDataHash = runtimeData.fromHash
-
-    // Unified will based the legend off ALL of the data maps received. Otherwise, it will use
-    let dataSet = obj.legend.unified ? obj.data : Object.values(runtimeData)
-
-    const colorDistributions = {
-      1: [1],
-      2: [1, 3],
-      3: [1, 3, 5],
-      4: [0, 2, 4, 6],
-      5: [0, 2, 4, 6, 7],
-      6: [0, 2, 3, 4, 5, 7],
-      7: [0, 2, 3, 4, 5, 6, 7],
-      8: [0, 2, 3, 4, 5, 6, 7, 8],
-      9: [0, 1, 2, 3, 4, 5, 6, 7, 8],
-      10: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
-    }
-
-    const applyColorToLegend = legendIdx => {
-      // Default to "bluegreen" color scheme if the passed color isn't valid
-      let mapColorPalette = obj.customColors || colorPalettes[obj.color] || colorPalettes['bluegreen']
-
-      // Handle Region Maps need for a 10th color
-      if (general.geoType === 'us-region' && mapColorPalette.length < 10 && mapColorPalette.length > 8) {
-        if (!general.palette.isReversed) {
-          mapColorPalette.push(chroma(mapColorPalette[8]).darken(0.75).hex())
-        } else {
-          mapColorPalette.unshift(chroma(mapColorPalette[0]).darken(0.75).hex())
-        }
-      }
-
-      let colorIdx = legendIdx - specialClasses
-
-      // Special Classes (No Data)
-      if (result[legendIdx].special) {
-        const specialClassColors = chroma.scale(['#D4D4D4', '#939393']).colors(specialClasses)
-
-        return specialClassColors[legendIdx]
-      }
-
-      if (obj.color.includes('qualitative')) return mapColorPalette[colorIdx]
-
-      let amt = Math.max(result.length - specialClasses, 1)
-      let distributionArray = colorDistributions[amt]
-
-      let specificColor
-      if (distributionArray) {
-        specificColor = distributionArray[colorIdx]
-      } else if (mapColorPalette[colorIdx]) {
-        specificColor = colorIdx
-      } else {
-        specificColor = mapColorPalette.length - 1
-      }
-
-      return mapColorPalette[specificColor]
-    }
-
-    let specialClasses = 0
-    let specialClassesHash = {}
-
-    // Special classes
-    if (obj.legend.specialClasses.length) {
-      if (typeof obj.legend.specialClasses[0] === 'object') {
-        obj.legend.specialClasses.forEach(specialClass => {
-          dataSet = dataSet.filter(row => {
-            const val = String(row[specialClass.key])
-
-            if (specialClass.value === val) {
-              if (undefined === specialClassesHash[val]) {
-                specialClassesHash[val] = true
-
-                result.push({
-                  special: true,
-                  value: val,
-                  label: specialClass.label
-                })
-
-                result[result.length - 1].color = applyColorToLegend(result.length - 1)
-
-                specialClasses += 1
-              }
-
-              let specialColor = ''
-
-              // color the state if val is in row
-              specialColor = result.findIndex(p => p.value === val)
-
-              newLegendMemo.set(hashObj(row), specialColor)
-
-              return false
-            }
-
-            return true
-          })
-        })
-      } else {
-        dataSet = dataSet.filter(row => {
-          const val = row[primaryCol]
-
-          if (obj.legend.specialClasses.includes(val)) {
-            // apply the special color to the legend
-            if (undefined === specialClassesHash[val]) {
-              specialClassesHash[val] = true
-
-              result.push({
-                special: true,
-                value: val
-              })
-
-              result[result.length - 1].color = applyColorToLegend(result.length - 1)
-
-              specialClasses += 1
-            }
-
-            let specialColor = ''
-
-            // color the state if val is in row
-            if (Object.values(row).includes(val)) {
-              specialColor = result.findIndex(p => p.value === val)
-            }
-
-            newLegendMemo.set(hashObj(row), specialColor)
-
-            return false
-          }
-
-          return true
-        })
-      }
-    }
-
-    // Category
-    if ('category' === type) {
-      let uniqueValues = new Map()
-      let count = 0
-
-      for (let i = 0; i < dataSet.length; i++) {
-        let row = dataSet[i]
-        let value = isBubble && categoricalCol && row[categoricalCol] ? row[categoricalCol] : row[primaryCol]
-        if (undefined === value) continue
-
-        if (false === uniqueValues.has(value)) {
-          uniqueValues.set(value, [hashObj(row)])
-          count++
-        } else {
-          uniqueValues.get(value).push(hashObj(row))
-        }
-
-        if (count === 10) break // Can only have 10 categorical items for now
-      }
-
-      let sorted = [...uniqueValues.keys()]
-
-      if (obj.legend.additionalCategories) {
-        obj.legend.additionalCategories.forEach(additionalCategory => {
-          if (additionalCategory && indexOfIgnoreType(sorted, additionalCategory) === -1) {
-            sorted.push(additionalCategory)
-          }
-        })
-      }
-
-      // Apply custom sorting or regular sorting
-      let configuredOrder = obj.legend.categoryValuesOrder ?? []
-
-      if (configuredOrder.length) {
-        sorted.sort((a, b) => {
-          let aVal = configuredOrder.indexOf(a)
-          let bVal = configuredOrder.indexOf(b)
-          if (aVal === bVal) return 0
-          if (aVal === -1) return 1
-          if (bVal === -1) return -1
-          return aVal - bVal
-        })
-      } else {
-        sorted.sort((a, b) => a - b)
-      }
-
-      // Add legend item for each
-      sorted.forEach(val => {
-        result.push({
-          value: val
-        })
-
-        let lastIdx = result.length - 1
-        let arr = uniqueValues.get(val)
-
-        if (arr) {
-          arr.forEach(hashedRow => newLegendMemo.set(hashedRow, lastIdx))
-        }
-      })
-
-      // Add color to new legend item
-      for (let i = 0; i < result.length; i++) {
-        result[i].color = applyColorToLegend(i)
-      }
-
-      legendMemo.current = newLegendMemo
-
-      // before returning the legend result
-      // add property for bin number and set to index location
-      result.forEach((row, i) => {
-        row.bin = i // set bin number to index
-      })
-
-      // Move all special legend items from "Special Classes"  to the end of the legend
-      if (config.legend.showSpecialClassesLast) {
-        let specialRows = result.filter(d => d.special === true)
-        let otherRows = result.filter(d => !d.special)
-        result = [...otherRows, ...specialRows]
-      }
-
-      return result
-    }
-
-    let uniqueValues = {}
-    dataSet.forEach(datum => {
-      uniqueValues[datum[primaryCol]] = true
-    })
-
-    let legendNumber = Math.min(number, Object.keys(uniqueValues).length)
-
-    // Separate zero
-    if (true === obj.legend.separateZero && !config.general.equalNumberOptIn) {
-      let addLegendItem = false
-
-      for (let i = 0; i < dataSet.length; i++) {
-        if (dataSet[i][primaryCol] === 0) {
-          addLegendItem = true
-
-          let row = dataSet.splice(i, 1)[0]
-
-          newLegendMemo.set(hashObj(row), result.length)
-          i--
-        }
-      }
-
-      if (addLegendItem) {
-        legendNumber -= 1 // This zero takes up one legend item
-
-        // Add new legend item
-        result.push({
-          min: 0,
-          max: 0
-        })
-
-        let lastIdx = result.length - 1
-
-        // Add color to new legend item
-        result[lastIdx].color = applyColorToLegend(lastIdx)
-      }
-    }
-
-    // Sort data for use in equalnumber or equalinterval
-    if (config.general.type !== 'us-geocode') {
-      dataSet = dataSet
-        .filter(row => typeof row[primaryCol] === 'number')
-        .sort((a, b) => {
-          let aNum = a[primaryCol]
-          let bNum = b[primaryCol]
-
-          return aNum - bNum
-        })
-    }
-
-    // Equal Number
-    if (type === 'equalnumber') {
-      // start work on changing legend functionality
-      // FALSE === ignore old version for now.
-      if (!config.general.equalNumberOptIn) {
-        let numberOfRows = dataSet.length
-
-        let remainder
-        let changingNumber = legendNumber
-
-        let chunkAmt
-
-        // Loop through the array until it has been split into equal subarrays
-        while (numberOfRows > 0) {
-          remainder = numberOfRows % changingNumber
-          chunkAmt = Math.floor(numberOfRows / changingNumber)
-
-          if (remainder > 0) {
-            chunkAmt += 1
-          }
-
-          let removedRows = dataSet.splice(0, chunkAmt)
-
-          let min = removedRows[0][primaryCol],
-            max = removedRows[removedRows.length - 1][primaryCol]
-
-          // eslint-disable-next-line
-          removedRows.forEach(row => {
-            newLegendMemo.set(hashObj(row), result.length)
-          })
-
-          result.push({
-            min,
-            max
-          })
-
-          result[result.length - 1].color = applyColorToLegend(result.length - 1)
-
-          changingNumber -= 1
-          numberOfRows -= chunkAmt
-        }
-      } else {
-        // get nums
-        let hasZeroInData = dataSet.filter(obj => obj[config.columns.primary.name] === 0).length > 0
-        let domainNums = new Set(dataSet.map(item => item[config.columns.primary.name]))
-
-        domainNums = d3.extent(domainNums)
-
-        let colors = colorPalettes[config.color]
-        let colorRange = colors.slice(0, config.legend.numberOfItems)
-
-        let scale = d3
-          .scaleQuantile()
-          .domain([...new Set(dataSet.map(item => Math.round(item[config.columns.primary.name])))]) // min/max values
-          .range(colorRange) // set range to our colors array
-
-        let breaks = scale.quantiles()
-
-        breaks = breaks.map(item => Math.round(item))
-
-        // if seperating zero force it into breaks
-        if (breaks[0] !== 0) {
-          breaks.unshift(0)
-        }
-
-        // eslint-disable-next-line array-callback-return
-        breaks.map((item, index) => {
-          const setMin = index => {
-            let min = breaks[index]
-
-            // if first break is a seperated zero, min is zero
-            if (index === 0 && config.legend.separateZero) {
-              min = 0
-            }
-
-            // if we're on the second break, and seperating out zero, increment min to 1.
-            if (index === 1 && config.legend.separateZero) {
-              min = 1
-            }
-
-            // // in starting position and zero in the data
-            // if((index === config.legend.specialClasses?.length ) && (config.legend.specialClasses.length !== 0)) {
-            //     min = breaks[index]
-            // }
-            return min
-          }
-
-          const setMax = (index, min) => {
-            let max = breaks[index + 1] - 1
-
-            // check if min and max range are the same
-            // if (min === max + 1) {
-            //     max = breaks[index + 1]
-            // }
-
-            if (index === 0 && config.legend.separateZero) {
-              max = 0
-            }
-            // if ((index === config.legend.specialClasses.length && config.legend.specialClasses.length !== 0) && !config.legend.separateZero && hasZeroInData) {
-            //     max = 0;
-            // }
-
-            if (index + 1 === breaks.length) {
-              max = domainNums[1]
-            }
-
-            return max
-          }
-
-          let min = setMin(index)
-          let max = setMax(index, min)
-
-          result.push({
-            min,
-            max,
-            color: scale(item)
-          })
-
-          dataSet.forEach((row, dataIndex) => {
-            let number = row[config.columns.primary.name]
-            let updated = 0
-
-            // check if we're seperating zero out
-            updated = config.legend.separateZero && hasZeroInData ? index : index
-            // check for special classes
-            updated = config.legend.specialClasses ? updated + config.legend.specialClasses.length : index
-
-            if (result[updated]?.min === (null || undefined) || result[updated]?.max === (null || undefined)) return
-
-            if (number >= result[updated].min && number <= result[updated].max) {
-              newLegendMemo.set(hashObj(row), updated)
-            }
-          })
-        })
-      }
-    }
-
-    // Equal Interval
-    if (type === 'equalinterval' && dataSet?.length !== 0) {
-      if (!dataSet || dataSet.length === 0) {
-        setState({
-          ...state,
-          runtime: {
-            ...config.runtime,
-            editorErrorMessage: 'Error setting equal interval legend type'
-          }
-        })
-        return
-      }
-      dataSet = dataSet.filter(row => row[primaryCol] !== undefined)
-      let dataMin = dataSet[0][primaryCol]
-      let dataMax = dataSet[dataSet.length - 1][primaryCol]
-
-      let pointer = 0 // Start at beginning of dataSet
-
-      for (let i = 0; i < legendNumber; i++) {
-        let interval = Math.abs(dataMax - dataMin) / legendNumber
-
-        let min = dataMin + interval * i
-        let max = min + interval
-
-        // If this is the last loop, assign actual max of data as the end point
-        if (i === legendNumber - 1) max = dataMax
-
-        // Add rows in dataSet that belong to this new legend item since we've got the data sorted
-        while (pointer < dataSet.length && dataSet[pointer][primaryCol] <= max) {
-          newLegendMemo.set(hashObj(dataSet[pointer]), result.length)
-          pointer += 1
-        }
-
-        let range = {
-          min: Math.round(min * 100) / 100,
-          max: Math.round(max * 100) / 100
-        }
-
-        result.push(range)
-
-        result[result.length - 1].color = applyColorToLegend(result.length - 1)
-      }
-    }
-
-    result.forEach((legendItem, idx) => {
-      legendItem.color = applyColorToLegend(idx, specialClasses, result)
-    })
-
-    legendMemo.current = newLegendMemo
-
-    //----------
-    // DEV-784
-    // before returning the legend result
-    // add property for bin number and set to index location
-    result.forEach((row, i) => {
-      row.bin = i // set bin number to index
-    })
-
-    // Move all special legend items from "Special Classes"  to the end of the legend
-    if (config.legend.showSpecialClassesLast) {
-      let specialRows = result.filter(d => d.special === true)
-      let otherRows = result.filter(d => !d.special)
-      result = [...otherRows, ...specialRows]
-    }
-    //-----------
-
-    return result
-  })
-
-  // eslint-disable-next-line
-  const generateRuntimeFilters = useCallback((obj, hash, runtimeFilters) => {
+  const generateRuntimeFilters = (obj, hash, runtimeFilters) => {
     if (typeof obj === 'undefined' || undefined === obj.filters || obj.filters.length === 0) return []
 
     let filters = []
 
-    if (hash) filters.fromHash = hash
+    if (hash) filters['fromHash'] = hash
 
     obj?.filters.forEach(({ columnName, label, labels, queryParameter, orderedValues, active, values, type, showDropdown }, idx) => {
       let newFilter = runtimeFilters[idx]
@@ -882,18 +371,17 @@ const CdcMap = (props: CdcMapProperties) => {
 
       if (type !== 'url') {
         values = getUniqueValues(config.data, columnName)
-
-        if (obj.filters[idx].order === 'asc') {
-          values = values.sort(sortAsc)
-        }
-
-        if (obj.filters[idx].order === 'desc') {
-          values = values.sort(sortDesc)
-        }
-
-        if (obj.filters[idx].order === 'cust') {
-          if (obj.filters[idx]?.values.length > 0) {
-            values = obj.filters[idx].values
+        switch (obj.filters[idx].order) {
+          case 'asc': {
+            values = values.sort(sortAsc)
+          }
+          case 'desc': {
+            values = values.sort(sortDesc)
+          }
+          case 'cust': {
+            if (obj.filters[idx]?.values.length > 0) {
+              values = obj.filters[idx].values
+            }
           }
         }
       } else {
@@ -921,11 +409,11 @@ const CdcMap = (props: CdcMapProperties) => {
     })
 
     return filters
-  })
+  }
 
   // Calculates what's going to be displayed on the map and data table at render.
   // eslint-disable-next-line
-  const generateRuntimeData = useCallback((obj, filters, hash, test) => {
+  const generateRuntimeData = (obj, filters?, hash?, test?) => {
     try {
       const result = {}
 
@@ -983,7 +471,7 @@ const CdcMap = (props: CdcMapProperties) => {
     } catch (e) {
       console.error('COVE: ', e) // eslint-disable-line
     }
-  })
+  }
 
   const outerContainerRef = useCallback(node => {
     if (node !== null) {
@@ -1048,7 +536,7 @@ const CdcMap = (props: CdcMapProperties) => {
       }
 
       // Check if it's a special value. If it is not, apply the designated prefix and suffix
-      if (false === config.legend.specialClasses.includes(String(value))) {
+      if (false === (config.legend.specialClasses as unknown as string[]).includes(String(value))) {
         formattedValue = (columnObj.prefix || '') + formattedValue + (columnObj.suffix || '')
       }
     }
@@ -1124,8 +612,6 @@ const CdcMap = (props: CdcMapProperties) => {
       delete legendItem.disabled
     })
 
-    newLegend.runtimeDataHash = runtimeLegend.runtimeDataHash
-
     dispatch({ type: 'SET_RUNTIME_LEGEND', payload: newLegend })
   }
 
@@ -1189,7 +675,7 @@ const CdcMap = (props: CdcMapProperties) => {
   // todo: convert to store or context eventually.
   const { buildTooltip } = useTooltip({ config, displayGeoName, displayDataAsText, supportedStatesFipsCodes })
 
-  const applyTooltipsToGeo = (geoName, row, returnType = 'string') => {
+  const applyTooltipsToGeo = (geoName?, row?, returnType = 'string') => {
     let toolTipText = buildTooltip(row, geoName, '')
 
     // We convert the markup into JSX and add a navigation link if it's going into a modal.
@@ -1198,7 +684,7 @@ const CdcMap = (props: CdcMapProperties) => {
 
       if (config.columns.hasOwnProperty('navigate') && row[config.columns.navigate.name]) {
         toolTipText.push(
-          // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions
+          // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-noninteractive-element-interactions
           <ul className='navigation-link' key='modal-navigation-link' onClick={() => navigationHandler(row[config.columns.navigate.name])}>
             {config.tooltips.linkLabel}
             <ExternalIcon className='inline-icon ml-1' />
@@ -1519,7 +1005,7 @@ const CdcMap = (props: CdcMapProperties) => {
     const hashFilters = hashObj(config.filters)
     let filters
 
-    if (config.filters && hashFilters !== runtimeFilters.fromHash) {
+    if (config.filters && hashFilters !== (runtimeFilters as unknown as { fromHash }).fromHash) {
       filters = generateRuntimeFilters(config, hashFilters, runtimeFilters)
 
       if (filters) {
@@ -1543,10 +1029,10 @@ const CdcMap = (props: CdcMapProperties) => {
     // Data
     if (hashData !== runtimeData.fromHash && config.data?.fromColumn) {
       const newRuntimeData = generateRuntimeData(config, filters || runtimeFilters, hashData)
-      dispatch({ type: 'SET_RUNTIME_DATA', payload: newRuntimeData })
+      setRuntimeData(newRuntimeData)
     } else {
-      if (hashLegend !== runtimeLegend.fromHash && undefined === runtimeData.init) {
-        const legend = generateRuntimeLegend(config, runtimeData, hashLegend)
+      if (hashLegend !== (runtimeLegend as unknown as { fromHash }).fromHash && undefined === runtimeData.init) {
+        const legend = generateRuntimeLegend(config, runtimeData, hashLegend, legendMemo, setErrorMessage)
         dispatch({ type: 'SET_RUNTIME_LEGEND', payload: legend })
       }
     }
@@ -1556,7 +1042,7 @@ const CdcMap = (props: CdcMapProperties) => {
     const hashLegend = generateRuntimeLegendHash()
 
     // Legend - Update when runtimeData does
-    const legend = generateRuntimeLegend(config, runtimeData, hashLegend)
+    const legend = generateRuntimeLegend(config, runtimeData, hashLegend, legendMemo, setErrorMessage)
     dispatch({ type: 'SET_RUNTIME_LEGEND', payload: legend })
   }, [runtimeData, config.legend.unified, config.legend.showSpecialClassesLast, config.legend.separateZero, config.general.equalNumberOptIn, config.legend.numberOfItems, config.legend.specialClasses]) // eslint-disable-line
 
@@ -1690,7 +1176,7 @@ const CdcMap = (props: CdcMapProperties) => {
             {title && (
               <header className={general.showTitle === true ? 'visible' : 'hidden'} {...(!general.showTitle || !config.general.title ? { 'aria-hidden': true } : { 'aria-hidden': false })}>
                 {/* eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex */}
-                <div role='heading' className={'map-title ' + general.headerColor} tabIndex='0' aria-level='2'>
+                <div role='heading' className={'map-title ' + general.headerColor} tabIndex={0} aria-level={2}>
                   {general.superTitle && <sup>{parse(general.superTitle)}</sup>}
                   <div>{parse(title)}</div>
                 </div>
@@ -1703,9 +1189,11 @@ const CdcMap = (props: CdcMapProperties) => {
 
             <div
               role='button'
-              tabIndex='0'
+              tabIndex={0}
               className={mapContainerClasses.join(' ')}
+              // eslint-disable-next-line jsx-a11y/no-static-element-interactions
               onClick={e => closeModal(e)}
+              // eslint-disable-next-line jsx-a11y/no-static-element-interactions
               onKeyDown={e => {
                 if (e.keyCode === 13) {
                   closeModal(e)
@@ -1717,7 +1205,7 @@ const CdcMap = (props: CdcMapProperties) => {
               </a>
 
               {/* eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex */}
-              <section className='outline-none geography-container' ref={mapSvg} tabIndex='0' style={{ width: '100%' }}>
+              <section className='outline-none geography-container' ref={mapSvg} tabIndex={0} style={{ width: '100%' }}>
                 {currentViewport && (
                   <>
                     {modal && <Modal content={modal} />}
