--- packages/map/src/context.jsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-
-const ConfigContext = createContext({})
-
-export default ConfigContext
