+++ packages/map/src/types/RuntimeLegend.ts
@@ -1,0 +1,1 @@
+export type RuntimeLegend = { disabled; bin; color; special }[]
